+++ packages/cldr-core/__tests__/api/general/names.test.ts
@@ -1,4 +1,6 @@
+import { AltType } from '@phensley/cldr-schema';
 import { generalApi, languageBundle, INTERNALS } from '../../_helpers';
+import { DisplayNameOptions } from '../../../src';
 
 test('languages', () => {
   let api = generalApi('en');
@@ -8,10 +10,16 @@ test('languages', () => {
   expect(api.getLanguageDisplayName('ko')).toEqual('Korean');
 
   api = generalApi('es');
-  expect(api.getLanguageDisplayName('en')).toEqual('inglés');
-  expect(api.getLanguageDisplayName('de')).toEqual('alemán');
-  expect(api.getLanguageDisplayName('fr')).toEqual('francés');
-  expect(api.getLanguageDisplayName('ko')).toEqual('coreano');
+  expect(api.getLanguageDisplayName('en')).toEqual('Inglés');
+  expect(api.getLanguageDisplayName('de')).toEqual('Alemán');
+  expect(api.getLanguageDisplayName('fr')).toEqual('Francés');
+  expect(api.getLanguageDisplayName('ko')).toEqual('Coreano');
+
+  const opts: DisplayNameOptions = { context: 'middle-of-text' };
+  expect(api.getLanguageDisplayName('en', opts)).toEqual('inglés');
+  expect(api.getLanguageDisplayName('de', opts)).toEqual('alemán');
+  expect(api.getLanguageDisplayName('fr', opts)).toEqual('francés');
+  expect(api.getLanguageDisplayName('ko', opts)).toEqual('coreano');
 
   api = generalApi('de');
   expect(api.getLanguageDisplayName('en')).toEqual('Englisch');
@@ -29,17 +37,17 @@ test('languages', () => {
 test('regions', () => {
   let api = generalApi('en');
   expect(api.getRegionDisplayName('US')).toEqual('United States');
-  expect(api.getRegionDisplayName('US', 'short')).toEqual('US');
-  expect(api.getRegionDisplayName('US', 'INVALID')).toEqual('United States');
+  expect(api.getRegionDisplayName('US', { type: 'short' })).toEqual('US');
+  expect(api.getRegionDisplayName('US', { type: 'INVALID' as AltType })).toEqual('United States');
 
   expect(api.getRegionDisplayName('ZZ')).toEqual('Unknown Region');
 
-  expect(api.getRegionDisplayName('DE', 'variant')).toEqual('Germany');
-  expect(api.getRegionDisplayName('DE', 'narrow')).toEqual('Germany');
+  expect(api.getRegionDisplayName('DE', { type: 'variant' })).toEqual('Germany');
+  expect(api.getRegionDisplayName('DE', { type: 'narrow' })).toEqual('Germany');
 
   api = generalApi('es');
   expect(api.getRegionDisplayName('US')).toEqual('Estados Unidos');
-  expect(api.getRegionDisplayName('US', 'short')).toEqual('EE. UU.');
+  expect(api.getRegionDisplayName('US', { type: 'short' })).toEqual('EE. UU.');
   expect(api.getRegionDisplayName('ZZ')).toEqual('Región desconocida');
 });
 
@@ -52,11 +60,18 @@ test('scripts', () => {
   expect(api.getScriptDisplayName('Zzzz')).toEqual('Unknown Script');
 
   api = generalApi('es');
-  expect(api.getScriptDisplayName('Latn')).toEqual('latino');
-  expect(api.getScriptDisplayName('Cyrl')).toEqual('cirílico');
-  expect(api.getScriptDisplayName('Hant')).toEqual('tradicional');
-  expect(api.getScriptDisplayName('Arab')).toEqual('árabe');
-  expect(api.getScriptDisplayName('Zzzz')).toEqual('alfabeto desconocido');
+  expect(api.getScriptDisplayName('Latn')).toEqual('Latino');
+  expect(api.getScriptDisplayName('Cyrl')).toEqual('Cirílico');
+  expect(api.getScriptDisplayName('Hant')).toEqual('Tradicional');
+  expect(api.getScriptDisplayName('Arab')).toEqual('Árabe');
+  expect(api.getScriptDisplayName('Zzzz')).toEqual('Alfabeto desconocido');
+
+  const opts: DisplayNameOptions = { context: 'middle-of-text' };
+  expect(api.getScriptDisplayName('Latn', opts)).toEqual('latino');
+  expect(api.getScriptDisplayName('Cyrl', opts)).toEqual('cirílico');
+  expect(api.getScriptDisplayName('Hant', opts)).toEqual('tradicional');
+  expect(api.getScriptDisplayName('Arab', opts)).toEqual('árabe');
+  expect(api.getScriptDisplayName('Zzzz', opts)).toEqual('alfabeto desconocido');
 });
 
 test('internals', () => {