+++ packages/cldr-core/src/common/general.ts
@@ -1,6 +1,17 @@
+import { AltType, ContextType } from '@phensley/cldr-schema';
 
 export type ListPatternType = 'and' | 'and-short' | 'or' | 'unit-long' | 'unit-narrow' | 'unit-short';
 
 export type MeasurementSystem = 'us' | 'uk' | 'metric';
 
 export type MeasurementCategory = 'temperature';
+
+/**
+ * @alpha
+ */
+export interface DisplayNameOptions {
+
+  type?: AltType;
+
+  context?: ContextType;
+}