+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNRELEASED]
+### Fixed/Changed
+ - Added `DisplayNameOptions` to methods for language, script, and region names.
+
 ## [0.15.3]
 ### Fixed/Changed
  - Fix Decimal compare against Infinity.